% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{import_QS_files}
\alias{import_QS_files}
\title{Import data}
\usage{
import_QS_files(
  paths,
  Ch1_is_mutation = TRUE,
  annotations = NULL,
  sample_annotations = NULL,
  merge_wells = "none",
  merge_files = FALSE
)
}
\arguments{
\item{paths}{Character vector. Specifies which files and/or directories to
load. For directories all \code{.csv} files within are loaded.}

\item{Ch1_is_mutation}{Logical. Control if Ch1 is used to measure presence of
mutant DNA (and Ch2 for wild type DNA). If this is FALSE Ch2 is used
instead (and Ch1 for wild type).}

\item{annotations}{Named list or \code{data.frame}. Annotations to be added
to all samples.}

\item{sample_annotations}{\code{data.frame}. Similar to \code{annotations},
but annotations are specific to each sample. This need to include a column
"Sample", which will be used for joining.}

\item{merge_wells}{String. Controls if wells from the same sample
("Sample") should be merged within a dataset (file). There are 4 options:
\itemize{
  \item \code{merge_wells="yes"}: Merge wells. Discards QS merged wells (e.g. "M01").
  \item \code{merge_wells="no"}: Do **not** merge wells. Discard QS merged wells (e.g. "M01").
  \item \code{merge_wells="qs"}: The merged wells from QuantaSoft (e.g. "M01") is used if these are present.
  \item \code{merge_wells="none"}: No merging is done.
}
Default if "none".}

\item{merge_files}{Logical. If this and \code{merge_wells} is TRUE, samples
across files are also merged. Default is FALSE.}
}
\value{
A \code{data.frame} with the columns:
  \item{FileName:}{Name of the file imported.}
  \item{Well:}{Well identifier.}
  \item{Sample:}{Sample name from QuantaSoft.}
  \item{Ch1TargetType:}{The target type for channel 1.}
  \item{Ch2TargetType:}{The target type for channel 2.}
  \item{Target:}{Target name from QuantaSoft.}
  \item{MutantOnlyDroplets:}{The count of mutant only droplets.}
  \item{WildtypeOnlyDroplets:}{The count of wild type only droplets.}
  \item{DoubleNegativeDroplets:}{The count of double negative droplets.}
  \item{DoublePositiveDroplets:}{The count of double positive droplets.}
  \item{TotalDroplets:}{The total number of droplets.}
  \item{MergedWells:}{If \code{merge_wells=TRUE} this is a string of the wells merged.
     Otherwise this will be NA.}
  \item{NumberOfMergedWells:}{The number of wells merged.}
}
\description{
Imports .csv-files generated by QuantaSoft.
}
\seealso{
\itemize{
  \item \code{\link{train_simple_ddpcr_model}}, \code{\link{test_tumor_sample_simple}}
  \item \code{\link{train_integrated_ddpcr_model}}, \code{\link{test_tumor_sample_integrated}}
}
}
