% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_droplet_counts}
\alias{simulate_droplet_counts}
\alias{simulate_molecule_counts}
\title{Simulating ddPCR data}
\usage{
simulate_droplet_counts(l, r, a, b, c, n_drops, n_samples = 1)

simulate_molecule_counts(l, r, a, b, c, n_drops)
}
\arguments{
\item{l}{Numeric value. Concentration of wild type DNA (molecules/droplet).}

\item{r}{Numeric value. Concentration of mutant DNA (molecules/droplet).}

\item{a, b, c}{Numeric values. Error rate parameters for the statistical model.}

\item{n_drops}{Integer value. The number of droplets to simulate in each sample.}

\item{n_samples}{Integer value. The number of samples to simulate.}
}
\value{
Simulated droplet/molecule counts as \code{data.frame}.
}
\description{
Functions for simulating data from the statistical model underlying the \code{castle} package.
}
\examples{
# Simulate droplet counts

# Single cancer negative sample (r = 0)
simulate_droplet_counts(1, 0, 0.01, 0.01, 0.01, 14000)

# Multiple cancer positive samples (r = 0.1)
simulate_droplet_counts(1, 0.1, 0.01, 0.01, 0.01, 14000, 10)


# Simulate molecule counts for a cancer positive samples (r = 0.1)
simulate_molecule_counts(1, 0.1, 0.01, 0.01, 0.01, 14000)
}
