% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sum_log_p}
\alias{sum_log_p}
\title{Calculate sum of probabilities in log-space}
\usage{
sum_log_p(lp_vec)
}
\arguments{
\item{lp_vec}{numeric vector}
}
\value{
numeric value
}
\description{
Calculate sum of probabilities in log-space
}
