% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrated_test.R
\name{test_tumor_sample_integrated}
\alias{test_tumor_sample_integrated}
\title{Test sample using integrated CASTLE model}
\usage{
test_tumor_sample_integrated(
  N_WT_only,
  N_M_only,
  N_d_neg,
  N_d_pos,
  abc_grid,
  abc_grid_train_log_lik,
  alpha = 0.05,
  include_r_confidence_interval = TRUE,
  include_l_confidence_interval = TRUE
)
}
\arguments{
\item{include_l_confidence_interval}{}
}
\value{

}
\description{
Test sample using integrated CASTLE model
}
