% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{import_QS_files}
\alias{import_QS_files}
\title{Import data form QuantaSoft}
\usage{
import_QS_files(
  paths,
  Ch1_is_mutation = TRUE,
  annotations = NULL,
  sample_annotations = NULL,
  merge_wells = FALSE,
  merge_files = FALSE,
  include_qs_concentrations = FALSE
)
}
\arguments{
\item{merge_wells}{Logical. Controls if wells from the sample sample
("Sample") should be merged within a dataset. If \code{merge_wells="qs"}
the merged wells from QuantaSoft (MXX) is used. Default if FALSE.}

\item{merge_files}{Logical. If this and \code{merge_wells} is TRUE, samples
across files are also merged. Default is FALSE.}

\item{include_qs_concentrations}{}
}
\value{

}
\description{
Import data form QuantaSoft
}
