% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrated_test.R
\name{test_tumor_sample_integrated}
\alias{test_tumor_sample_integrated}
\title{Test sample using integrated CASTLE model}
\usage{
test_tumor_sample_integrated(
  test_samples,
  integrated_model,
  alpha = 0.05,
  include_r_CI = TRUE,
  include_l_CI = TRUE
)
}
\arguments{
\item{test_samples}{\code{data.frame} with samples to be tested  as rows. At
least the columns \code{N_WT_only}, \code{N_M_only}, \code{N_d_neg} and
\code{N_d_pos} should be present.}

\item{integrated_model}{A fitted model as generated by
\code{train_integrated_ddpcr_model}.}

\item{alpha}{The alpha used for the statistical test and confidence
intervals. Default is 0.05.}

\item{include_r_CI}{logical. Should confidence intervals for \code{r} be
included? Default is \code{TRUE}.}

\item{include_l_CI}{logical. Should confidence intervals for \code{l} be
included? Default is \code{TRUE}.}
}
\value{
A \code{data.frame} with results from the CASTLE algorithm (parameter
  estimates, CIs, p-values etc.) joined to the right of \code{test_samples}.
}
\description{
Tests samples for the presence of mutant DNA using the
  integrated version of the CASTLE algorithm.
}
\seealso{
\code{\link{train_integrated_ddpcr_model}}
}
