% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-simple.R
\name{test_tumor_sample_simple}
\alias{test_tumor_sample_simple}
\title{Test sample using simple CASTLE model}
\usage{
test_tumor_sample_simple(test_samples, simple_model, alpha = 0.05)
}
\arguments{
\item{test_samples}{\code{data.frame} with samples to be tested  as rows. At
least the columns \code{N_WT_only}, \code{N_M_only}, \code{N_d_neg} and
\code{N_d_pos} should be present.}

\item{simple_model}{A fitted model as generated by
\code{train_simple_ddpcr_model}.}

\item{alpha}{The alpha used for the statistical test and confidence
intervals. Default is 0.05.}

\item{include_r_CI}{logical. Should confidence intervals for \code{r} be
included? Default is \code{TRUE}.}

\item{include_l_CI}{logical. Should confidence intervals for \code{l} be
included? Default is \code{TRUE}.}
}
\value{
A \code{data.frame} with results from the CASTLE algorithm (parameter
  estimates, CIs, p-values etc.) joined to the right of \code{test_samples}.
}
\description{
Tests samples for the presence of mutant DNA using the
  simple version of the CASTLE algorithm.
}
\seealso{
\code{\link{train_simple_ddpcr_model}}
}
