% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/castle_simple.R
\name{test_tumor_sample_simple}
\alias{test_tumor_sample_simple}
\title{Test samples using simple CASTLE model}
\usage{
test_tumor_sample_simple(
  test_samples,
  simple_model,
  include_wildtype_CI = TRUE,
  include_mutant_CI = TRUE,
  alpha = 0.01
)
}
\arguments{
\item{test_samples}{\code{data.frame} with samples to be tested as rows. At
  least the following columns should be present:
    \itemize{
      \item{\code{WildtypeOnlyDroplets}}
      \item{\code{MutantOnlyDroplets}}
      \item{\code{DoubleNegativeDroplets}}
      \item{\code{DoublePositiveDroplets}}
   }
If data from QuataSoft is available, these can be imported using
\code{\link{import_QS_files}}.}

\item{simple_model}{A fitted model as generated by
\code{\link{train_simple_ddpcr_model}}.}

\item{include_wildtype_CI}{Logical. Should confidence intervals for wild type
DNA statistics be included? Default is \code{TRUE}.}

\item{include_mutant_CI}{Logical. Should confidence intervals for mutant
DNA statistics be included? Default is \code{TRUE}.}

\item{alpha}{The alpha used for the statistical test and confidence
intervals. Default is 0.01.}
}
\value{
A \code{data.frame} with results from the CASTLE algorithm (parameter
  estimates, CIs, p-values etc.) joined to the right of \code{test_samples}.
}
\description{
Tests samples for the presence of mutant DNA using the
  simple version of the CASTLE algorithm.
}
\seealso{
\code{\link{train_simple_ddpcr_model}}
}
