% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/castle_integrated.R
\name{train_integrated_ddpcr_model}
\alias{train_integrated_ddpcr_model}
\title{Train integrated CASTLE model}
\usage{
train_integrated_ddpcr_model(background_samples, abc_grid_resolution = 25)
}
\arguments{
\item{background_samples}{\code{data.frame} with cancer negative training
  samples as rows. At least the following columns should be present:
    \itemize{
      \item{\code{WildtypeOnlyDroplets}}
      \item{\code{MutantOnlyDroplets}}
      \item{\code{DoubleNegativeDroplets}}
      \item{\code{DoublePositiveDroplets}}
   }
If data from QuataSoft is available, these can be imported using
\code{\link{import_QS_files}}.}

\item{abc_grid_resolution}{The resolution of the 3D-grid on which the
integrated LR-test statistic is approximated. Default is 25 (equal to a
grid of 25^3 = 15,625 points).}
}
\value{
List of parameter estimates for the integrated CASLTE model. This
  should be used as input for \code{\link{test_tumor_sample_integrated}}.
}
\description{
Estimates the parameters of the statistical model for the
  integrated version the CASTLE algorithm based on a set of training samples,
  assumed to have no mutant DNA present.
}
\details{
In the test an integrated version of the Likelihood Ratio test
  statistic is used, and this is approximated using a Riemann sum on a
  3D-grid. The parameter \code{abc_grid_resolution} controls this
  approximation as the sum is calculated on \code{abc_grid_resolution^3}
  points. Higher values give better approximation but longer computation
  times.
}
\seealso{
\code{\link{import_QS_files}, \link{test_tumor_sample_integrated}}
}
