% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrated_test.R
\name{train_integrated_ddpcr_model}
\alias{train_integrated_ddpcr_model}
\title{Train integrated CASTLE model}
\usage{
train_integrated_ddpcr_model(
  N_WT_only_vec,
  N_M_only_vec,
  N_d_neg_vec,
  N_d_pos_vec,
  abc_grid_resolution = 25
)
}
\arguments{
\item{abc_grid_resolution}{}
}
\value{

}
\description{
Train integrated CASTLE model
}
