% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/castle_simple.R
\name{train_simple_ddpcr_model}
\alias{train_simple_ddpcr_model}
\title{Train simple CASTLE model}
\usage{
train_simple_ddpcr_model(background_samples)
}
\arguments{
\item{background_samples}{\code{data.frame} with cancer negative training
samples as rows. These are the samples used to train/estimate the
parameters of the model. At least the following columns should be present:
\itemize{
  \item{\code{WildtypeOnlyDroplets}}
  \item{\code{MutantOnlyDroplets}}
  \item{\code{DoubleNegativeDroplets}}
  \item{\code{DoublePositiveDroplets}} } If data from QuantaSoft is available,
these can be imported using \code{\link{import_QS_files}}.}
}
\value{
List of parameter estimates for the simple CASLTE model. This should
  be used as input for \code{\link{test_tumor_sample_simple}}.
}
\description{
Estimates the parameters of the statistical model for the simple
  version the CASTLE algorithm based on a set of training samples, assumed to
  have no mutant DNA present.
}
\seealso{
\code{\link{import_QS_files}},
  \code{\link{test_tumor_sample_simple}}
}
